requirements.txt
langchain-groq==0.2.4
langchain-core>=0.3.33,<0.4.0
langgraph==0.2.60
chromadb==0.6.3
sentence-transformers==3.3.1
streamlit==1.41.1
ragas==0.2.6
datasets==3.2.0
python-dotenv==1.0.1
